% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_button_response}
\alias{html_button_response}
\title{Create an HTML stimulus trial with button press response}
\usage{
html_button_response(stimulus, choices, prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE, button_html = NULL,
  margin_vertical = "0px", margin_horizontal = "8px")
}
\arguments{
\item{stimulus}{the HTML stimulus}

\item{choices}{character vector specifying labels}

\item{prompt}{the prompt to be displayed with the stimulus}

\item{stimulus_duration}{how long the stimulus is displayed}

\item{trial_duration}{how long the trial lasts}

\item{response_ends_trial}{default = TRUE}

\item{button_html}{HTML specifying the buttons (defaults to jsPsych default)}

\item{margin_vertical}{vertical margin of the buttons}

\item{margin_horizontal}{horizontal margin of the buttons}
}
\description{
Create an HTML stimulus trial with button press response
}
