% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_html.R
\name{html_keyboard_response}
\alias{html_keyboard_response}
\title{HTML stimulus, keyboard response}
\usage{
html_keyboard_response(stimulus, choices = any_key(), prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE)
}
\arguments{
\item{stimulus}{HTML stimulus}

\item{choices}{Character vector specifying the allowed response keys (default = any_key())}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}
}
\description{
HTML stimulus, keyboard response
}
