% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_html.R
\name{html_slider_response}
\alias{html_slider_response}
\title{HTML stimulus, slider response}
\usage{
html_slider_response(stimulus, labels = NULL,
  button_label = "Continue", min = 0, max = 100, start = 50,
  step = 1, slider_width = NULL, require_movement = FALSE,
  prompt = NULL, stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE)
}
\arguments{
\item{stimulus}{HTML stimulus}

\item{labels}{Character vector of labels, to be spaced equally along slider}

\item{button_label}{Text displayed on button at the end (default = "Continue")}

\item{min}{Minimum value of the slider (default = 0)}

\item{max}{Maximum value of the slider (default = 100)}

\item{start}{Starting value of the slider (default = 50)}

\item{step}{Smallest increment of slider movement (default = 1)}

\item{slider_width}{Horizontal width of slider (defaults to max width)}

\item{require_movement}{Does user need to move the slider to continue? (default = FALSE)}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}
}
\description{
HTML stimulus, slider response
}
