% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_image.R
\name{image_keyboard_response}
\alias{image_keyboard_response}
\title{Image stimulus, keyboard response}
\usage{
image_keyboard_response(stimulus, stimulus_height = NULL,
  stimulus_width = NULL, maintain_aspect_ratio = TRUE,
  choices = any_key(), prompt = NULL, stimulus_duration = NULL,
  trial_duration = NULL, margin_vertical = "0px",
  margin_horizontal = "8px", response_ends_trial = TRUE)
}
\arguments{
\item{stimulus}{Image file, specified using the resource() function}

\item{stimulus_height}{Image height in pixels (defaults to natural height)}

\item{stimulus_width}{Image width in pixels (defaults to natural width)}

\item{maintain_aspect_ratio}{Maintain aspect ratio if only one of height/width specified? (default = TRUE)}

\item{choices}{Character vector specifying the allowed response keys (default = any_key())}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}
}
\description{
Image stimulus, keyboard response
}
