% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_types.R
\name{js_string}
\alias{js_string}
\title{Specify js string directly}
\usage{
js_string(...)
}
\arguments{
\item{...}{arguments passed to the trial plugin}

\item{type}{the type of trial block}
}
\description{
Specify js string directly
}
