% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{jspsych_add}
\alias{jspsych_add}
\title{Add trial that shows an image and respond with keyboard}
\usage{
jspsych_add(xpt, type, options = list())
}
\arguments{
\item{xpt}{the experiment object}

\item{stimulus}{path to an image file}

\item{choices}{numeric codes for allowed responses}

\item{prompt}{text of the response prompt}
}
\description{
Add trial that shows an image and respond with keyboard
}
