% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{jspsych_create}
\alias{jspsych_create}
\title{Initialise an experiment}
\usage{
jspsych_create(files)
}
\arguments{
\item{files}{character listing files to include}
}
\description{
Initialise an experiment
}
