% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{jspsych_write}
\alias{jspsych_write}
\title{Convert the experiment object to HTML file}
\usage{
jspsych_write(xpt, path)
}
\arguments{
\item{xpt}{the experiment object}

\item{path}{where to write the experiment}
}
\description{
Convert the experiment object to HTML file
}
