% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_misc.R
\name{any_key}
\alias{any_key}
\alias{no_key}
\title{Keyboard responses}
\usage{
any_key()

no_key()
}
\description{
Keyboard responses
}
\details{
The \code{any_key()} function allows the user to continue to the
next trial by pressing any key. The \code{no_key()} function ensures that
no keypress can terminate the trial, nor will any response be recorded.
}
