% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_write.R
\name{make_experiment}
\alias{make_experiment}
\title{Make the experiment}
\usage{
make_experiment(timeline, path, resources, ...)
}
\arguments{
\item{timeline}{tl}

\item{path}{p}

\item{resources}{r}

\item{...}{pass to init}
}
\description{
Make the experiment
}
