% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{question_likert}
\alias{question_likert}
\title{Define a Likert question}
\usage{
question_likert(prompt, labels, required = FALSE, name = NULL)
}
\arguments{
\item{prompt}{The question}

\item{labels}{The set of labels to attach to the Likert scale}

\item{required}{Is this a required question (default = FALSE)}

\item{name}{A name for the question (if NULL, defaults to Q0, Q1, etc)}
}
\description{
Define a Likert question
}
