% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{question_multi_choice}
\alias{question_multi_choice}
\title{Define a multiple choice question}
\usage{
question_multi_choice(prompt, options, required = FALSE,
  horizontal = FALSE, name = NULL)
}
\arguments{
\item{prompt}{The question}

\item{options}{The set of possible answers}

\item{required}{Is this a required question (default = FALSE)}

\item{horizontal}{Should the options be laid out horizontally (default = FALSE)}

\item{name}{A name for the question (if NULL, defaults to Q0, Q1, etc)}
}
\description{
Define a multiple choice question
}
