% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{question_text}
\alias{question_text}
\title{Define a free text response question}
\usage{
question_text(prompt, placeholder = "", required = FALSE, rows = 1,
  columns = 40, name = NULL)
}
\arguments{
\item{prompt}{The question}

\item{placeholder}{The placeholder text in the response field (default = "")}

\item{required}{Is this a required question (default = FALSE)}

\item{rows}{How many text rows should the response box use? (default = 1)}

\item{columns}{How many text columns should the response box use? (default = 40)}

\item{name}{A name for the question (if NULL, defaults to Q0, Q1, etc)}
}
\description{
Define a free text response question
}
