% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responders.R
\name{respond_button}
\alias{respond_button}
\title{Define response buttons}
\usage{
respond_button(choices, button_html = NULL, margin_vertical = "0px",
  margin_horizontal = "8px")
}
\arguments{
\item{choices}{Character vector with labels for the buttons}

\item{button_html}{HTML templay specifying the buttons (defaults to jsPsych default)}

\item{margin_vertical}{Vertical margin of the buttons (default "0px")}

\item{margin_horizontal}{Horizontal margin of the buttons (default "8px")}
}
\description{
Define response buttons
}
