% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responders.R
\name{respond_key}
\alias{respond_key}
\title{Define a keyboard responder}
\usage{
respond_key(...)
}
\arguments{
\item{...}{Either single string "any" (default), "none", or characters specifying keys}
}
\description{
Define a keyboard responder
}
