% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responders.R
\name{respond_slider}
\alias{respond_slider}
\title{Define a response slider}
\usage{
respond_slider(labels = NULL, button_label = "Continue", min = 0,
  max = 100, start = 50, step = 1, slider_width = NULL,
  require_movement = FALSE)
}
\arguments{
\item{labels}{Character vector of labels, to be spaced equally along slider}

\item{button_label}{Text displayed on button at the end (default = "Continue")}

\item{min}{Minimum value of the slider (default = 0)}

\item{max}{Maximum value of the slider (default = 100)}

\item{start}{Starting value of the slider (default = 50)}

\item{step}{Smallest increment of slider movement (default = 1)}

\item{slider_width}{Horizontal width of slider (defaults to max width)}

\item{require_movement}{Does user need to move the slider to continue? (default = FALSE)}
}
\description{
Define a response slider
}
