% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_build.R
\name{run_locally}
\alias{run_locally}
\title{Run a jspsych experiment}
\usage{
run_locally(path = ".", xprmntr_host = "127.0.0.1",
  xprmntr_port = 8000)
}
\arguments{
\item{path}{path}

\item{xprmntr_host}{host defaults to 127.0.0.1}

\item{xprmntr_port}{port defaults to 8000}
}
\description{
Run a jspsych experiment
}
