% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stimuli.R
\name{stimulus_image}
\alias{stimulus_image}
\title{Define an image stimulus}
\usage{
stimulus_image(stimulus, stimulus_height = NULL, stimulus_width = NULL,
  maintain_aspect_ratio = TRUE)
}
\arguments{
\item{stimulus}{Image file, specified using the resource() function}

\item{stimulus_height}{Image height in pixels (defaults to natural height)}

\item{stimulus_width}{Image width in pixels (defaults to natural width)}

\item{maintain_aspect_ratio}{Maintain aspect ratio if only one of height/width specified? (default = TRUE)}
}
\description{
Define an image stimulus
}
