% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{string}
\alias{string}
\title{Specify js string concatenation directly}
\usage{
string(...)
}
\arguments{
\item{...}{arguments passed to the trial plugin}

\item{type}{the type of trial block}
}
\description{
Specify js string concatenation directly
}
