% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_likert}
\alias{survey_likert}
\title{Create a survey trial with Likert items}
\usage{
survey_likert(questions, shuffle = FALSE, preamble = "",
  scale_width = NULL, button_label = "Continue")
}
\arguments{
\item{questions}{A list of questions generated by survey_question()}

\item{shuffle}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{scale_width}{Width of the response scale in pixels (NULL defaults to max)}

\item{button_label}{Text shown in the button to end the trial (default = "Continue")}
}
\description{
Create a survey trial with Likert items
}
