% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_multi_choice}
\alias{survey_multi_choice}
\title{Create a survey trial with multiple choice items}
\usage{
survey_multi_choice(questions, shuffle = FALSE, preamble = "",
  button_label = "Continue")
}
\arguments{
\item{questions}{A list of questions generated by survey_question()}

\item{shuffle}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{button_label}{Text shown in the button to end the trial (default = "Continue")}
}
\description{
Create a survey trial with multiple choice items
}
