% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{survey_questions}
\alias{survey_questions}
\alias{question_likert}
\alias{question_multi}
\alias{question_text}
\title{Survey questions}
\usage{
question_likert(prompt, labels, required = FALSE, name = NULL)

question_multi(prompt, options, required = FALSE, horizontal = FALSE,
  name = NULL)

question_text(prompt, placeholder = "", required = FALSE, rows = 1,
  columns = 40, name = NULL)
}
\arguments{
\item{prompt}{The question to be asked}

\item{labels}{The set of labels to attach to the Likert scale}

\item{required}{Is this a required question? (default = FALSE)}

\item{name}{A name for the question (if NULL, defaults to Q0, Q1, etc)}

\item{options}{The set of possible answers to a multiple choice question}

\item{horizontal}{Should multiple choice options be laid out horizontally (default = FALSE)}

\item{placeholder}{Placeholder text for a free text response question (default = "")}

\item{rows}{How many text rows should a free text response box use? (default = 1)}

\item{columns}{How many text columns should should a free text response box use? (default = 40)}
}
\description{
Survey questions
}
