% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_build.R
\name{tl_add}
\alias{tl_add}
\title{Adds an arbitrary trial block to the timeline}
\usage{
tl_add(timeline, type, ...)
}
\arguments{
\item{timeline}{the timeline object}

\item{type}{the type of trial block}

\item{...}{arguments passed to the trial plugin}
}
\description{
Adds an arbitrary trial block to the timeline
}
