% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_build.R
\name{tl_new}
\alias{tl_new}
\title{Initialise a timeline}
\usage{
tl_new(scripts = "jspsych.js", stylesheets = "jspsych.css",
  resources = character())
}
\arguments{
\item{scripts}{character vector listing user files to include}

\item{stylesheets}{character vector listing jspsych files to include}

\item{resources}{dsgdsf}
}
\description{
Initialise a timeline
}
\details{
If no arguments are specified, the resources field is empty,
and the jspsych field contains the core jspsych library and css files
}
