% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_build.R
\name{trial}
\alias{trial}
\title{Adds an arbitrary trial block to the timeline}
\usage{
trial(type, ...)
}
\arguments{
\item{type}{the type of trial block}

\item{...}{arguments passed to the trial plugin}
}
\description{
Adds an arbitrary trial block to the timeline
}
