% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_html.R
\name{trial_html_button}
\alias{trial_html_button}
\title{Trial with HTML stimulus and button response}
\usage{
trial_html_button(stimulus, choices, button_html = NULL,
  margin_vertical = "0px", margin_horizontal = "8px", prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE, ...)
}
\arguments{
\item{stimulus}{HTML stimulus}

\item{choices}{Character vector with labels for the buttons}

\item{button_html}{HTML templay specifying the buttons (defaults to jsPsych default)}

\item{margin_vertical}{Vertical margin of the buttons (default "0px")}

\item{margin_horizontal}{Horizontal margin of the buttons (default "8px")}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}

\item{...}{Arguments to be passed to trial()}
}
\description{
Trial with HTML stimulus and button response
}
