% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_html.R
\name{trial_html_key}
\alias{trial_html_key}
\title{Trial with HTML stimulus and keyboard response}
\usage{
trial_html_key(stimulus, response = respond_key("any"), prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE, ...)
}
\arguments{
\item{stimulus}{A stimulus object generated by stimulus_html()}

\item{response}{A response object generated by respond_key()}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}

\item{...}{Arguments to be passed to trial()}
}
\description{
Trial with HTML stimulus and keyboard response
}
