% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_image.R
\name{trial_image_button}
\alias{trial_image_button}
\title{Trial with image stimulus and button response}
\usage{
trial_image_button(stimulus, response, prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE, ...)
}
\arguments{
\item{stimulus}{A stimulus object generated by stimulus_image()}

\item{response}{A response object generated by respond_button()}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}

\item{...}{Arguments to be passed to trial()}
}
\description{
Trial with image stimulus and button response
}
