% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_image.R
\name{trial_image_key}
\alias{trial_image_key}
\title{Trial with image stimulus and keyboard response}
\usage{
trial_image_key(stimulus, stimulus_height = NULL,
  stimulus_width = NULL, maintain_aspect_ratio = TRUE,
  choices = any_key(), prompt = NULL, stimulus_duration = NULL,
  trial_duration = NULL, response_ends_trial = TRUE, ...)
}
\arguments{
\item{stimulus}{Image file, specified using the resource() function}

\item{stimulus_height}{Image height in pixels (defaults to natural height)}

\item{stimulus_width}{Image width in pixels (defaults to natural width)}

\item{maintain_aspect_ratio}{Maintain aspect ratio if only one of height/width specified? (default = TRUE)}

\item{choices}{A character vector, any_key(), or no_key()}

\item{prompt}{The prompt to be displayed with the stimulus (can include HTML markup)}

\item{stimulus_duration}{How long to display stimulus in milliseconds? (default is to remain until response)}

\item{trial_duration}{How long to wait for response in milliseconds? (default is to wait indefinitely)}

\item{response_ends_trial}{Does the trial end with response (default = TRUE) or continue until trial duration reached?}

\item{...}{Arguments to be passed to trial()}
}
\description{
Trial with image stimulus and keyboard response
}
