% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{trial_survey_likert}
\alias{trial_survey_likert}
\title{Survey trial with Likert items}
\usage{
trial_survey_likert(questions, randomize_question_order = FALSE,
  preamble = "", scale_width = NULL, button_label = "Continue")
}
\arguments{
\item{questions}{A list of questions, each generated by question_likert()}

\item{randomize_question_order}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{scale_width}{Width of the response scale in pixels (NULL defaults to max)}

\item{button_label}{Text displayed on button at the end (default = "Continue")}
}
\description{
Survey trial with Likert items
}
