% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{trial_survey_multi_select}
\alias{trial_survey_multi_select}
\title{Survey trial with multiple selection items}
\usage{
trial_survey_multi_select(questions, randomize_question_order = FALSE,
  preamble = "", button_label = "Continue",
  required_message = "You must choose at least one response for this question")
}
\arguments{
\item{questions}{A list of questions, each generated by question_multi_choice()}

\item{randomize_question_order}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{button_label}{Text displayed on button at the end (default = "Continue")}

\item{required_message}{Text warning shown when a user does not select an option}
}
\description{
Survey trial with multiple selection items
}
