% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{trial_survey_text}
\alias{trial_survey_text}
\title{Survey trial with free text responses}
\usage{
trial_survey_text(questions, randomize_question_order = FALSE,
  preamble = "", button_label = "Continue")
}
\arguments{
\item{questions}{A list of questions, each generated by question_text()}

\item{randomize_question_order}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{button_label}{Text displayed on button at the end (default = "Continue")}
}
\description{
Survey trial with free text responses
}
