% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_build.R
\name{use_variable}
\alias{use_variable}
\title{Insert a timeline variable}
\usage{
use_variable(name)
}
\arguments{
\item{name}{name of the variable to insert}
}
\description{
Insert a timeline variable
}
