% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_build.R
\name{with_variable}
\alias{with_variable}
\title{Attach a timeline variable to timeline object}
\usage{
with_variable(timeline, name, values)
}
\arguments{
\item{timeline}{the timeline object}

\item{name}{name of the variable}

\item{values}{possible values for the variable}
}
\description{
Attach a timeline variable to timeline object
}
