% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xpt_image_keyboard_response}
\alias{xpt_image_keyboard_response}
\title{Add trial that shows an image and respond with keyboard}
\usage{
xpt_image_keyboard_response(xpt, stimulus, choices, prompt)
}
\arguments{
\item{xpt}{the experiment object}

\item{stimulus}{path to an image file}

\item{choices}{numeric codes for allowed responses}

\item{prompt}{text of the response prompt}
}
\description{
Add trial that shows an image and respond with keyboard
}
