% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{xpt_write}
\alias{xpt_write}
\title{Convert the experiment object to HTML file}
\usage{
xpt_write(xpt, file)
}
\arguments{
\item{xpt}{the experiment object}
}
\description{
Convert the experiment object to HTML file
}
