% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_keyboard_response}
\alias{html_keyboard_response}
\title{Create an HTML stimulus trial with keyboard response}
\usage{
html_keyboard_response(stimulus, choices = any_key(), prompt = NULL,
  stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE)
}
\arguments{
\item{stimulus}{the HTML stimulus}

\item{choices}{character vector specifying allowed response keys}

\item{prompt}{the prompt to be displayed with the stimulus}

\item{stimulus_duration}{how long the stimulus is displayed}

\item{trial_duration}{how long the trial lasts}

\item{response_ends_trial}{default = TRUE}
}
\description{
Create an HTML stimulus trial with keyboard response
}
