% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_slider_response}
\alias{html_slider_response}
\title{Create an HTML stimulus trial with slider response}
\usage{
html_slider_response(stimulus, labels = NULL,
  button_label = "Continue", min = 0, max = 100, start = 50,
  step = 1, slider_width = NULL, require_movement = FALSE,
  prompt = NULL, stimulus_duration = NULL, trial_duration = NULL,
  response_ends_trial = TRUE)
}
\arguments{
\item{stimulus}{the HTML stimulus}

\item{labels}{labels to be spaced equally along slider}

\item{button_label}{text for the button}

\item{min}{min value of slider}

\item{max}{max value of slider}

\item{start}{start point of slider}

\item{step}{smallest increment of slider movement}

\item{slider_width}{horizontal width of slider}

\item{require_movement}{does user need to move the slider?}

\item{prompt}{the prompt to be displayed with the stimulus}

\item{stimulus_duration}{how long the stimulus is displayed}

\item{trial_duration}{how long the trial lasts}

\item{response_ends_trial}{default = TRUE}
}
\description{
Create an HTML stimulus trial with slider response
}
