% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{jspsych_add}
\alias{jspsych_add}
\title{Add trial that shows an image and respond with keyboard}
\usage{
jspsych_add(xpt, type, ...)
}
\arguments{
\item{xpt}{the experiment object}

\item{type}{the type of trial}

\item{...}{arguments passed to the trial plugin}
}
\description{
Add trial that shows an image and respond with keyboard
}
