% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_write.R
\name{jspsych_init}
\alias{jspsych_init}
\title{Options on initialisation}
\usage{
jspsych_init(timeline, ...)
}
\arguments{
\item{timeline}{a timeline object}

\item{...}{arguments passed to init}
}
\description{
Options on initialisation
}
