% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{jspsych_init}
\alias{jspsych_init}
\title{Options on initialisation}
\usage{
jspsych_init(xpt, ...)
}
\arguments{
\item{xpt}{the experiment}

\item{...}{arguments passed to init}
}
\description{
Options on initialisation
}
