% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{survey_likert}
\alias{survey_likert}
\title{Survey trial, Likert items}
\usage{
survey_likert(questions, shuffle = FALSE, preamble = "",
  scale_width = NULL, button_label = "Continue")
}
\arguments{
\item{questions}{A list of questions, each generated by question_likert()}

\item{shuffle}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{scale_width}{Width of the response scale in pixels (NULL defaults to max)}

\item{button_label}{Text displayed on button at the end (default = "Continue")}
}
\description{
Survey trial, Likert items
}
