% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_survey.R
\name{survey_multi_choice}
\alias{survey_multi_choice}
\title{Survey trial, multiple choice items}
\usage{
survey_multi_choice(questions, shuffle = FALSE, preamble = "",
  button_label = "Continue")
}
\arguments{
\item{questions}{A list of questions, each generated by question_multi_choice()}

\item{shuffle}{Should the question order be randomised? (default = FALSE)}

\item{preamble}{Text to display above the survey items}

\item{button_label}{Text displayed on button at the end (default = "Continue")}
}
\description{
Survey trial, multiple choice items
}
